refactor: move useDashboardState into own file
